+++ NEWS.md
@@ -4,6 +4,7 @@
 * `if_all.()`/`if_any.()`
   + No longer try to use `.by` columns (#225)
   + Can use `n()` in filtering (#226)
+* `pivot_longer.()`: No longer converts factor values to character (#202)
 
 # tidytable 0.6.0